+++ identity/src/peer_id.rs
@@ -18,7 +18,6 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use multiaddr::{Multiaddr, Protocol};
 use multihash::{Code, Error, Multihash};
 use rand::Rng;
 use std::{convert::TryFrom, fmt, str::FromStr};
@@ -89,17 +88,6 @@ impl PeerId {
             }
             _ => Err(multihash),
         }
-    }
-
-    /// Tries to extract a [`PeerId`] from the given [`Multiaddr`].
-    ///
-    /// In case the given [`Multiaddr`] ends with `/p2p/<peer-id>`, this function
-    /// will return the encapsulated [`PeerId`], otherwise it will return `None`.
-    pub fn try_from_multiaddr(address: &Multiaddr) -> Option<PeerId> {
-        address.iter().last().and_then(|p| match p {
-            Protocol::P2p(hash) => PeerId::from_multihash(hash).ok(),
-            _ => None,
-        })
     }
 
     /// Generates a random peer ID from a cryptographically secure PRNG.
@@ -287,33 +275,5 @@ mod tests {
             let peer_id = PeerId::random();
             assert_eq!(peer_id, PeerId::from_bytes(&peer_id.to_bytes()).unwrap());
         }
-    }
-
-    #[test]
-    fn extract_peer_id_from_multi_address() {
-        let address = "/memory/1234/p2p/12D3KooWGQmdpzHXCqLno4mMxWXKNFQHASBeF99gTm2JR8Vu5Bdc"
-            .to_string()
-            .parse()
-            .unwrap();
-
-        #[allow(deprecated)]
-        let peer_id = PeerId::try_from_multiaddr(&address).unwrap();
-
-        assert_eq!(
-            peer_id,
-            "12D3KooWGQmdpzHXCqLno4mMxWXKNFQHASBeF99gTm2JR8Vu5Bdc"
-                .parse()
-                .unwrap()
-        );
-    }
-
-    #[test]
-    fn no_panic_on_extract_peer_id_from_multi_address_if_not_present() {
-        let address = "/memory/1234".to_string().parse().unwrap();
-
-        #[allow(deprecated)]
-        let maybe_empty = PeerId::try_from_multiaddr(&address);
-
-        assert!(maybe_empty.is_none());
     }
 }